backend/requirements.txt
Django
django-cors-headers
django-ratelimit
django-redis
djangorestframework
djangorestframework-jwt
django-filter
gunicorn
whitenoise
sqlalchemy
mysqlclient
drf_yasg